+++ data/types/settings/pump/basal_rate_start_test.go
@@ -6,72 +6,14 @@ import (
 	. "github.com/onsi/gomega"
 
 	dataNormalizer "github.com/tidepool-org/platform/data/normalizer"
-	dataTypesBasalTest "github.com/tidepool-org/platform/data/types/basal/test"
 	"github.com/tidepool-org/platform/data/types/settings/pump"
+	dataTypesSettingsPumpTest "github.com/tidepool-org/platform/data/types/settings/pump/test"
 	dataTypesTest "github.com/tidepool-org/platform/data/types/test"
 	errorsTest "github.com/tidepool-org/platform/errors/test"
 	"github.com/tidepool-org/platform/pointer"
 	"github.com/tidepool-org/platform/structure"
 	structureValidator "github.com/tidepool-org/platform/structure/validator"
-	"github.com/tidepool-org/platform/test"
 )
-
-func NewBasalRateStart(startMinimum int) *pump.BasalRateStart {
-	datum := pump.NewBasalRateStart()
-	datum.Rate = pointer.FromFloat64(test.RandomFloat64FromRange(pump.BasalRateStartRateMinimum, pump.BasalRateStartRateMaximum))
-	if startMinimum == pump.BasalRateStartStartMinimum {
-		datum.Start = pointer.FromInt(pump.BasalRateStartStartMinimum)
-	} else {
-		datum.Start = pointer.FromInt(test.RandomIntFromRange(startMinimum, pump.BasalRateStartStartMaximum))
-	}
-	return datum
-}
-
-func CloneBasalRateStart(datum *pump.BasalRateStart) *pump.BasalRateStart {
-	if datum == nil {
-		return nil
-	}
-	clone := pump.NewBasalRateStart()
-	clone.Rate = pointer.CloneFloat64(datum.Rate)
-	clone.Start = pointer.CloneInt(datum.Start)
-	return clone
-}
-
-func NewBasalRateStartArray() *pump.BasalRateStartArray {
-	datum := pump.NewBasalRateStartArray()
-	*datum = append(*datum, NewBasalRateStart(pump.BasalRateStartStartMinimum))
-	*datum = append(*datum, NewBasalRateStart(*datum.Last().Start+1))
-	*datum = append(*datum, NewBasalRateStart(*datum.Last().Start+1))
-	return datum
-}
-
-func CloneBasalRateStartArray(datumArray *pump.BasalRateStartArray) *pump.BasalRateStartArray {
-	if datumArray == nil {
-		return nil
-	}
-	clone := pump.NewBasalRateStartArray()
-	for _, datum := range *datumArray {
-		*clone = append(*clone, CloneBasalRateStart(datum))
-	}
-	return clone
-}
-
-func NewBasalRateStartArrayMap() *pump.BasalRateStartArrayMap {
-	datum := pump.NewBasalRateStartArrayMap()
-	datum.Set(dataTypesBasalTest.NewScheduleName(), NewBasalRateStartArray())
-	return datum
-}
-
-func CloneBasalRateStartArrayMap(datumArrayMap *pump.BasalRateStartArrayMap) *pump.BasalRateStartArrayMap {
-	if datumArrayMap == nil {
-		return nil
-	}
-	clone := pump.NewBasalRateStartArrayMap()
-	for datumName, datumArray := range *datumArrayMap {
-		clone.Set(datumName, CloneBasalRateStartArray(datumArray))
-	}
-	return clone
-}
 
 var _ = Describe("BasalRateStart", func() {
 	It("BasalRateStartRateMaximum is expected", func() {
@@ -108,7 +50,7 @@ var _ = Describe("BasalRateStart", func() {
 		Context("Validate", func() {
 			DescribeTable("validates the datum",
 				func(mutator func(datum *pump.BasalRateStart), expectedErrors ...error) {
-					datum := NewBasalRateStart(pump.BasalRateStartStartMinimum)
+					datum := dataTypesSettingsPumpTest.NewBasalRateStart(pump.BasalRateStartStartMinimum)
 					mutator(datum)
 					dataTypesTest.ValidateWithExpectedOrigins(structureValidator.NewValidatableWithIntAdapter(datum, pointer.FromInt(pump.BasalRateStartStartMinimum)), structure.Origins(), expectedErrors...)
 				},
@@ -149,7 +91,7 @@ var _ = Describe("BasalRateStart", func() {
 
 			DescribeTable("validates the datum with minimum start",
 				func(mutator func(datum *pump.BasalRateStart), expectedErrors ...error) {
-					datum := NewBasalRateStart(pump.BasalRateStartStartMinimum)
+					datum := dataTypesSettingsPumpTest.NewBasalRateStart(pump.BasalRateStartStartMinimum)
 					mutator(datum)
 					dataTypesTest.ValidateWithExpectedOrigins(structureValidator.NewValidatableWithIntAdapter(datum, pointer.FromInt(pump.BasalRateStartStartMinimum)), structure.Origins(), expectedErrors...)
 				},
@@ -168,7 +110,7 @@ var _ = Describe("BasalRateStart", func() {
 
 			DescribeTable("validates the datum with non-minimum start",
 				func(mutator func(datum *pump.BasalRateStart), expectedErrors ...error) {
-					datum := NewBasalRateStart(pump.BasalRateStartStartMinimum + 1)
+					datum := dataTypesSettingsPumpTest.NewBasalRateStart(pump.BasalRateStartStartMinimum + 1)
 					mutator(datum)
 					dataTypesTest.ValidateWithExpectedOrigins(structureValidator.NewValidatableWithIntAdapter(datum, pointer.FromInt(pump.BasalRateStartStartMinimum+1)), structure.Origins(), expectedErrors...)
 				},
@@ -193,9 +135,9 @@ var _ = Describe("BasalRateStart", func() {
 			DescribeTable("normalizes the datum",
 				func(mutator func(datum *pump.BasalRateStart)) {
 					for _, origin := range structure.Origins() {
-						datum := NewBasalRateStart(pump.BasalRateStartStartMinimum + 1)
+						datum := dataTypesSettingsPumpTest.NewBasalRateStart(pump.BasalRateStartStartMinimum + 1)
 						mutator(datum)
-						expectedDatum := CloneBasalRateStart(datum)
+						expectedDatum := dataTypesSettingsPumpTest.CloneBasalRateStart(datum)
 						normalizer := dataNormalizer.New()
 						Expect(normalizer).ToNot(BeNil())
 						datum.Normalize(normalizer.WithOrigin(origin))
@@ -251,7 +193,7 @@ var _ = Describe("BasalRateStart", func() {
 				),
 				Entry("single invalid",
 					func(datum *pump.BasalRateStartArray) {
-						invalid := NewBasalRateStart(pump.BasalRateStartStartMinimum)
+						invalid := dataTypesSettingsPumpTest.NewBasalRateStart(pump.BasalRateStartStartMinimum)
 						invalid.Rate = nil
 						*datum = append(*datum, invalid)
 					},
@@ -259,32 +201,32 @@ var _ = Describe("BasalRateStart", func() {
 				),
 				Entry("single valid",
 					func(datum *pump.BasalRateStartArray) {
-						*datum = append(*datum, NewBasalRateStart(pump.BasalRateStartStartMinimum))
+						*datum = append(*datum, dataTypesSettingsPumpTest.NewBasalRateStart(pump.BasalRateStartStartMinimum))
 					},
 				),
 				Entry("multiple invalid",
 					func(datum *pump.BasalRateStartArray) {
-						*datum = append(*datum, NewBasalRateStart(pump.BasalRateStartStartMinimum))
-						invalid := NewBasalRateStart(*datum.Last().Start + 1)
+						*datum = append(*datum, dataTypesSettingsPumpTest.NewBasalRateStart(pump.BasalRateStartStartMinimum))
+						invalid := dataTypesSettingsPumpTest.NewBasalRateStart(*datum.Last().Start + 1)
 						invalid.Rate = nil
 						*datum = append(*datum, invalid)
-						*datum = append(*datum, NewBasalRateStart(*datum.Last().Start+1))
+						*datum = append(*datum, dataTypesSettingsPumpTest.NewBasalRateStart(*datum.Last().Start+1))
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotExists(), "/1/rate"),
 				),
 				Entry("multiple valid",
 					func(datum *pump.BasalRateStartArray) {
-						*datum = append(*datum, NewBasalRateStart(pump.BasalRateStartStartMinimum))
-						*datum = append(*datum, NewBasalRateStart(*datum.Last().Start+1))
-						*datum = append(*datum, NewBasalRateStart(*datum.Last().Start+1))
+						*datum = append(*datum, dataTypesSettingsPumpTest.NewBasalRateStart(pump.BasalRateStartStartMinimum))
+						*datum = append(*datum, dataTypesSettingsPumpTest.NewBasalRateStart(*datum.Last().Start+1))
+						*datum = append(*datum, dataTypesSettingsPumpTest.NewBasalRateStart(*datum.Last().Start+1))
 					},
 				),
 				Entry("multiple errors",
 					func(datum *pump.BasalRateStartArray) {
-						invalid := NewBasalRateStart(pump.BasalRateStartStartMinimum)
+						invalid := dataTypesSettingsPumpTest.NewBasalRateStart(pump.BasalRateStartStartMinimum)
 						invalid.Rate = nil
 						*datum = append(*datum, nil, invalid)
-						*datum = append(*datum, nil, NewBasalRateStart(*datum.Last().Start+1))
+						*datum = append(*datum, nil, dataTypesSettingsPumpTest.NewBasalRateStart(*datum.Last().Start+1))
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotExists(), "/0"),
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotExists(), "/1/rate"),
@@ -297,9 +239,9 @@ var _ = Describe("BasalRateStart", func() {
 			DescribeTable("normalizes the datum",
 				func(mutator func(datum *pump.BasalRateStartArray)) {
 					for _, origin := range structure.Origins() {
-						datum := NewBasalRateStartArray()
+						datum := dataTypesSettingsPumpTest.NewBasalRateStartArray()
 						mutator(datum)
-						expectedDatum := CloneBasalRateStartArray(datum)
+						expectedDatum := dataTypesSettingsPumpTest.CloneBasalRateStartArray(datum)
 						normalizer := dataNormalizer.New()
 						Expect(normalizer).ToNot(BeNil())
 						datum.Normalize(normalizer.WithOrigin(origin))
@@ -332,14 +274,14 @@ var _ = Describe("BasalRateStart", func() {
 			})
 
 			It("returns the first element if the array has one element", func() {
-				*datum = append(*datum, NewBasalRateStart(pump.BasalRateStartStartMinimum))
+				*datum = append(*datum, dataTypesSettingsPumpTest.NewBasalRateStart(pump.BasalRateStartStartMinimum))
 				Expect(datum.First()).To(Equal((*datum)[0]))
 			})
 
 			It("returns the first element if the array has multiple elements", func() {
-				*datum = append(*datum, NewBasalRateStart(pump.BasalRateStartStartMinimum))
-				*datum = append(*datum, NewBasalRateStart(*datum.Last().Start+1))
-				*datum = append(*datum, NewBasalRateStart(*datum.Last().Start+1))
+				*datum = append(*datum, dataTypesSettingsPumpTest.NewBasalRateStart(pump.BasalRateStartStartMinimum))
+				*datum = append(*datum, dataTypesSettingsPumpTest.NewBasalRateStart(*datum.Last().Start+1))
+				*datum = append(*datum, dataTypesSettingsPumpTest.NewBasalRateStart(*datum.Last().Start+1))
 				Expect(datum.First()).To(Equal((*datum)[0]))
 			})
 		})
@@ -356,14 +298,14 @@ var _ = Describe("BasalRateStart", func() {
 			})
 
 			It("returns the last element if the array has one element", func() {
-				*datum = append(*datum, NewBasalRateStart(pump.BasalRateStartStartMinimum))
+				*datum = append(*datum, dataTypesSettingsPumpTest.NewBasalRateStart(pump.BasalRateStartStartMinimum))
 				Expect(datum.Last()).To(Equal((*datum)[0]))
 			})
 
 			It("returns the last element if the array has multiple elements", func() {
-				*datum = append(*datum, NewBasalRateStart(pump.BasalRateStartStartMinimum))
-				*datum = append(*datum, NewBasalRateStart(*datum.Last().Start+1))
-				*datum = append(*datum, NewBasalRateStart(*datum.Last().Start+1))
+				*datum = append(*datum, dataTypesSettingsPumpTest.NewBasalRateStart(pump.BasalRateStartStartMinimum))
+				*datum = append(*datum, dataTypesSettingsPumpTest.NewBasalRateStart(*datum.Last().Start+1))
+				*datum = append(*datum, dataTypesSettingsPumpTest.NewBasalRateStart(*datum.Last().Start+1))
 				Expect(datum.Last()).To(Equal((*datum)[2]))
 			})
 		})
@@ -398,7 +340,9 @@ var _ = Describe("BasalRateStart", func() {
 					func(datum *pump.BasalRateStartArrayMap) { *datum = *pump.NewBasalRateStartArrayMap() },
 				),
 				Entry("empty name",
-					func(datum *pump.BasalRateStartArrayMap) { datum.Set("", NewBasalRateStartArray()) },
+					func(datum *pump.BasalRateStartArrayMap) {
+						datum.Set("", dataTypesSettingsPumpTest.NewBasalRateStartArray())
+					},
 				),
 				Entry("nil value",
 					func(datum *pump.BasalRateStartArrayMap) { datum.Set("", nil) },
@@ -406,7 +350,7 @@ var _ = Describe("BasalRateStart", func() {
 				),
 				Entry("single invalid",
 					func(datum *pump.BasalRateStartArrayMap) {
-						invalid := NewBasalRateStartArray()
+						invalid := dataTypesSettingsPumpTest.NewBasalRateStartArray()
 						(*invalid)[0].Start = nil
 						datum.Set("one", invalid)
 					},
@@ -414,33 +358,33 @@ var _ = Describe("BasalRateStart", func() {
 				),
 				Entry("single valid",
 					func(datum *pump.BasalRateStartArrayMap) {
-						datum.Set("one", NewBasalRateStartArray())
+						datum.Set("one", dataTypesSettingsPumpTest.NewBasalRateStartArray())
 					},
 				),
 				Entry("multiple invalid",
 					func(datum *pump.BasalRateStartArrayMap) {
-						invalid := NewBasalRateStartArray()
+						invalid := dataTypesSettingsPumpTest.NewBasalRateStartArray()
 						(*invalid)[0].Start = nil
-						datum.Set("one", NewBasalRateStartArray())
+						datum.Set("one", dataTypesSettingsPumpTest.NewBasalRateStartArray())
 						datum.Set("two", invalid)
-						datum.Set("three", NewBasalRateStartArray())
+						datum.Set("three", dataTypesSettingsPumpTest.NewBasalRateStartArray())
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotExists(), "/two/0/start"),
 				),
 				Entry("multiple valid",
 					func(datum *pump.BasalRateStartArrayMap) {
-						datum.Set("one", NewBasalRateStartArray())
-						datum.Set("two", NewBasalRateStartArray())
-						datum.Set("three", NewBasalRateStartArray())
+						datum.Set("one", dataTypesSettingsPumpTest.NewBasalRateStartArray())
+						datum.Set("two", dataTypesSettingsPumpTest.NewBasalRateStartArray())
+						datum.Set("three", dataTypesSettingsPumpTest.NewBasalRateStartArray())
 					},
 				),
 				Entry("multiple errors",
 					func(datum *pump.BasalRateStartArrayMap) {
-						invalid := NewBasalRateStartArray()
+						invalid := dataTypesSettingsPumpTest.NewBasalRateStartArray()
 						(*invalid)[0].Start = nil
 						datum.Set("one", nil)
 						datum.Set("two", invalid)
-						datum.Set("three", NewBasalRateStartArray())
+						datum.Set("three", dataTypesSettingsPumpTest.NewBasalRateStartArray())
 					},
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotExists(), "/one"),
 					errorsTest.WithPointerSource(structureValidator.ErrorValueNotExists(), "/two/0/start"),
@@ -452,9 +396,9 @@ var _ = Describe("BasalRateStart", func() {
 			DescribeTable("normalizes the datum",
 				func(mutator func(datum *pump.BasalRateStartArrayMap)) {
 					for _, origin := range structure.Origins() {
-						datum := NewBasalRateStartArrayMap()
+						datum := dataTypesSettingsPumpTest.NewBasalRateStartArrayMap()
 						mutator(datum)
-						expectedDatum := CloneBasalRateStartArrayMap(datum)
+						expectedDatum := dataTypesSettingsPumpTest.CloneBasalRateStartArrayMap(datum)
 						normalizer := dataNormalizer.New()
 						Expect(normalizer).ToNot(BeNil())
 						datum.Normalize(normalizer.WithOrigin(origin))